app/requirements.txt
fastapi
uvicorn
python-multipart
onnxruntime
numpy
Pillow
psycopg2-binary